refactor(redskilled): separate GitHub publication transport Refs #3887
