+++ .changeset/redskilled-durable-github-publication.md
@@ -1,0 +1,5 @@
+---
+"@reddb-io/redskilled": minor
+---
+
+Route Project-scoped GitHub pushes, pull requests, issues, and comments through a daemon-owned durable ACP outbox with idempotent response-loss reconciliation and no Worker-held credentials.
